Lab_6/requirements.txt
Flask==3.1.2
pytest==9.0.1
prometheus-flask-exporter==0.23.1
psycopg2-binary==2.9.10
